Move async management to async header.
